Move fetch to separate file
