+++ src/server/browserContext.ts
@@ -16,8 +16,6 @@
  */
 
 import * as os from 'os';
-import * as url from 'url';
-import fetch from 'node-fetch';
 import { TimeoutSettings } from '../utils/timeoutSettings';
 import { debugMode, mkdirIfNeeded, createGuid } from '../utils/utils';
 import { Browser, BrowserOptions } from './browser';
@@ -36,8 +34,6 @@ import { Tracing } from './trace/recorder/tracing';
 import { HarTracer } from './supplements/har/harTracer';
 import { RecorderSupplement } from './supplements/recorderSupplement';
 import * as consoleApiSource from '../generated/consoleApiSource';
-
-const HttpsProxyAgent = require('https-proxy-agent');
 
 export abstract class BrowserContext extends SdkObject {
   static Events = {
@@ -181,70 +177,6 @@ export abstract class BrowserContext extends SdkObject {
     const binding = new PageBinding(name, playwrightBinding, needsHandle, world);
     this._pageBindings.set(identifier, binding);
     await this._doExposeBinding(binding);
-  }
-
-  async fetch(params: types.FetchOptions): Promise<{fetchResponse?: types.FetchResponse, error?: string}> {
-    try {
-      const cookies = await this.cookies(params.url);
-      const valueArray = cookies.map(c => `${c.name}=${c.value}`);
-      const clientCookie = params.headers?.['cookie'];
-      if (clientCookie)
-        valueArray.unshift(clientCookie);
-      const cookieHeader = valueArray.join('; ');
-      if (cookieHeader) {
-        if (!params.headers)
-          params.headers = {};
-        params.headers['cookie'] = cookieHeader;
-      }
-      if (!params.method)
-        params.method = 'GET';
-      let agent;
-      if (this._options.proxy) {
-        // TODO: support bypass proxy
-        const proxyOpts = url.parse(this._options.proxy.server);
-        if (this._options.proxy.username)
-          proxyOpts.auth = `${this._options.proxy.username}:${this._options.proxy.password || ''}`;
-        agent = new HttpsProxyAgent(proxyOpts);
-      }
-
-      // TODO: set user agent
-      const response = await fetch(params.url, {
-        method: params.method,
-        headers: params.headers,
-        body: params.postData,
-        agent});
-      const body = await response.buffer();
-      const setCookies = response.headers.raw()['set-cookie'];
-      if (setCookies) {
-        const cookies: types.SetNetworkCookieParam[] = [];
-        for (const header of setCookies) {
-          // Decode cookie value?
-          const cookie: types.SetNetworkCookieParam | null = parseCookie(header);
-          if (!cookie)
-            continue;
-          if (!cookie.domain)
-            cookie.url = params.url;
-          cookies.push(cookie);
-        }
-        if (cookies.length)
-          await this.addCookies(cookies);
-      }
-
-      const headers: types.HeadersArray = [];
-      for (const [name, value] of response.headers.entries())
-        headers.push({ name, value });
-      return {
-        fetchResponse: {
-          url: response.url,
-          status: response.status,
-          statusText: response.statusText,
-          headers,
-          body
-        }
-      };
-    } catch (e) {
-      return { error: String(e) };
-    }
   }
 
   async grantPermissions(permissions: string[], origin?: string) {
@@ -528,49 +460,3 @@ export function normalizeProxySettings(proxy: types.ProxySettings): types.ProxyS
     bypass = bypass.split(',').map(t => t.trim()).join(',');
   return { ...proxy, server, bypass };
 }
-
-function parseCookie(header: string) {
-  const pairs = header.split(';').filter(s => s.trim().length > 0).map(p => p.split('=').map(s => s.trim()));
-  if (!pairs.length)
-    return null;
-  const [name, value] = pairs[0];
-  const cookie: types.NetworkCookie = {
-    name,
-    value,
-    domain: '',
-    path: '',
-    expires: -1,
-    httpOnly: false,
-    secure: false,
-    sameSite: 'Lax' // None for non-chromium
-  };
-  for (let i = 1; i < pairs.length; i++) {
-    const [name, value] = pairs[i];
-    switch (name.toLowerCase()) {
-      case 'expires':
-        const expiresMs = (+new Date(value));
-        if (isFinite(expiresMs))
-          cookie.expires = expiresMs / 1000;
-        break;
-      case 'max-age':
-        const maxAgeSec = parseInt(value, 10);
-        if (isFinite(maxAgeSec))
-          cookie.expires = Date.now() / 1000 + maxAgeSec;
-        break;
-      case 'domain':
-        cookie.domain = value || '';
-        break;
-      case 'path':
-        cookie.path = value || '';
-        break;
-      case 'secure':
-        cookie.secure = true;
-        break;
-      case 'httponly':
-        cookie.httpOnly = true;
-        break;
-    }
-  }
-  return cookie;
-}
-
